Remove usage of nodemon
Kinda difficult to decide here but ultimately decided to remove it because this code will mostly be used in Glitch.

In Glitch, the app automatically restarts every time you make a change anyways, making this redundant. Plus, when running with good ol' node, it'll actually say that there's an error in the Tools button rather than spinning infinitely.