Add auth back into every endpoint
If you try out the unauthenticated endpoints from the API explorer, they don't show up in the logs since they're unauthenticated (obviously)!

It'd be nice for there to be a way to specify headers via JWT that show up in the explorer for every endpoint, but in the meantime we'll roll with this!